separate out the common.ts file into parent folder
This was a common source of merge conflicts, and was also just getting long.

TEST=build
