+++ studio/studio-plugin/createStudioPlugin.ts
@@ -5,7 +5,7 @@ import parsePageFile from './ts-morph/parsePageFile'
 import configureServer from './configureServer'
 import { StudioProps } from '../client/components/Studio'
 import getRootPath from './getRootPath'
-import { getSourceFile } from './common/common'
+import { getSourceFile } from './common'
 import { moduleNameToComponentMetadata } from './componentMetadata'
 import getPagePath from './getPagePath'
 import openBrowser from 'react-dev-utils/openBrowser.js'
