+++ studio/studio-plugin/ts-morph/parseComponentMetadata.test.ts
@@ -1,5 +1,5 @@
 import getRootPath from '../getRootPath'
-import { getSourceFile } from '../common/common'
+import { getSourceFile } from '../common'
 import parseComponentMetadata from './parseComponentMetadata'
 import { PropTypes } from '../../types'
