+++ studio/studio-plugin/ts-morph/parseImports.ts
@@ -1,5 +1,5 @@
 import { SourceFile, SyntaxKind } from 'ts-morph'
-import { getSourceFile } from '../common/common'
+import { getSourceFile } from '../common'
 
 /**
  * Returns a mapping of import identifier (i.e. path or module name)
