+++ studio/studio-plugin/ts-morph/parseInitialProps.ts
@@ -1,6 +1,6 @@
 import { SourceFile, ts } from 'ts-morph'
 import { PropState } from '../../shared/models'
-import parseObjectLiteralExpression from '../common/parseObjectLiteralExpression'
+import { parseObjectLiteralExpression } from '../common'
 
 export default function parseInitialProps(sourceFile: SourceFile): PropState {
   const initialPropsSymbol = sourceFile.getExportSymbols().find(s => s.compilerSymbol.escapedName === 'initialProps')
