+++ studio/studio-plugin/ts-morph/parsePageFile.ts
@@ -1,57 +1,10 @@
 import { JsxAttribute, ts, SourceFile, JsxElement, JsxFragment } from 'ts-morph'
 import { ComponentState, PossibleModuleNames, PageState } from '../../shared/models'
-import { getComponentName, getComponentNodes, getDefaultExport, getJsxAttributeValue, getPropName, getSourceFile } from '../common/common'
+import { getComponentName, getComponentNodes, getDefaultExport, getJsxAttributeValue, getPropName, validatePropState, getSourceFile } from '../common'
 import { v1 } from 'uuid'
 import parseImports from './parseImports'
 import { moduleNameToComponentMetadata } from '../componentMetadata'
-import validatePropState from '../common/validatePropState'
 
-function parseLayoutState(
-  sourceFile: SourceFile,
-  imports: Record<string, string[]>
-): { layoutState: ComponentState, layoutNode: JsxElement | JsxFragment } {
-  const defaultExport = getDefaultExport(sourceFile)
-  const returnStatement = defaultExport.getFirstDescendantByKind(ts.SyntaxKind.ReturnStatement)
-  if (!returnStatement) {
-    throw new Error('No return statement found for page')
-  }
-  const JsxNodeWrapper = returnStatement.getFirstChildByKind(ts.SyntaxKind.ParenthesizedExpression)
-    ?? returnStatement
-  const topLevelJsxNode = JsxNodeWrapper.getChildren()
-    .find(n => n.getKind() === ts.SyntaxKind.JsxElement || n.getKind() === ts.SyntaxKind.JsxFragment)
-  if (!topLevelJsxNode) {
-    throw new Error('Unable to find top level JSX element or JsxFragment type from file.')
-  }
-
-  let layoutState: ComponentState
-  if (topLevelJsxNode.getKind() === ts.SyntaxKind.JsxElement) {
-    const name = getComponentName((topLevelJsxNode as JsxElement).getOpeningElement())
-    layoutState = {
-      name,
-      props: {},
-      uuid: v1(),
-      moduleName: 'builtIn'
-    }
-    const isBuiltinJsxElement = name.charAt(0) === name.charAt(0).toLowerCase()
-    if (!isBuiltinJsxElement && !['Fragment', 'React.Fragment'].includes(name)) {
-      layoutState.moduleName = getComponentModuleName(name, imports, true)
-    }
-  } else {
-    // This handles the React.Fragment shorthand <></>
-    layoutState = {
-      name: '',
-      props: {},
-      uuid: v1(),
-      moduleName: 'builtIn'
-    }
-  }
-
-  return {
-    layoutState,
-    layoutNode: topLevelJsxNode as JsxElement | JsxFragment
-  }
-}
-// TODO(oshi): move parsePageFile to the top of this file
 export default function parsePageFile(filePath: string): PageState {
   const sourceFile = getSourceFile(filePath)
   const imports = parseImports(sourceFile)
@@ -101,6 +54,52 @@ export default function parsePageFile(filePath: string): PageState {
   return {
     layoutState,
     componentsState
+  }
+}
+
+function parseLayoutState(
+  sourceFile: SourceFile,
+  imports: Record<string, string[]>
+): { layoutState: ComponentState, layoutNode: JsxElement | JsxFragment } {
+  const defaultExport = getDefaultExport(sourceFile)
+  const returnStatement = defaultExport.getFirstDescendantByKind(ts.SyntaxKind.ReturnStatement)
+  if (!returnStatement) {
+    throw new Error('No return statement found for page')
+  }
+  const JsxNodeWrapper = returnStatement.getFirstChildByKind(ts.SyntaxKind.ParenthesizedExpression)
+    ?? returnStatement
+  const topLevelJsxNode = JsxNodeWrapper.getChildren()
+    .find(n => n.getKind() === ts.SyntaxKind.JsxElement || n.getKind() === ts.SyntaxKind.JsxFragment)
+  if (!topLevelJsxNode) {
+    throw new Error('Unable to find top level JSX element or JsxFragment type from file.')
+  }
+
+  let layoutState: ComponentState
+  if (topLevelJsxNode.getKind() === ts.SyntaxKind.JsxElement) {
+    const name = getComponentName((topLevelJsxNode as JsxElement).getOpeningElement())
+    layoutState = {
+      name,
+      props: {},
+      uuid: v1(),
+      moduleName: 'builtIn'
+    }
+    const isBuiltinJsxElement = name.charAt(0) === name.charAt(0).toLowerCase()
+    if (!isBuiltinJsxElement && !['Fragment', 'React.Fragment'].includes(name)) {
+      layoutState.moduleName = getComponentModuleName(name, imports, true)
+    }
+  } else {
+    // This handles the React.Fragment shorthand <></>
+    layoutState = {
+      name: '',
+      props: {},
+      uuid: v1(),
+      moduleName: 'builtIn'
+    }
+  }
+
+  return {
+    layoutState,
+    layoutNode: topLevelJsxNode as JsxElement | JsxFragment
   }
 }
